# server/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is API documentation for Redland Museum's DIGS administration portal",
    "version": "1.0.0",
    "title": "Swagger Redland Museum",
    "contact": {
      "email": "tyronewessnolasco@gmail.com"
    }
  },
  "host": "rm.kongroo.xyz",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/artefacts": {
      "get": {
        "tags": [
          "artefacts"
        ],
        "summary": "List all artefacts",
        "description": "",
        "operationId": "listArtefacts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "404": {
            "description": "Artefacts not found"
          }
        }
      },
      "post": {
        "tags": [
          "artefacts"
        ],
        "summary": "Add a new artefact to the list",
        "description": "",
        "operationId": "addArtefacts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Artefact that needs to be added to the list",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Artefact"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/artefacts/{id}": {
      "get": {
        "tags": [
          "artefacts"
        ],
        "summary": "Find artefact by ID",
        "description": "Returns a single artefact",
        "operationId": "getArtefactById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of artefact to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Artefact"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Artefact not found"
          }
        }
      },
      "put": {
        "tags": [
          "artefacts"
        ],
        "summary": "Updates a artefact with form data",
        "description": "Updates artefact",
        "operationId": "updateArtefactWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "ID of the to be updated artefact",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the artefact",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "Updated description of the artefact",
            "required": false,
            "type": "string"
          },
          {
            "name": "acquisitionDate",
            "in": "formData",
            "description": "Updated acquisition date of the artefact",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "createdAt",
            "in": "formData",
            "description": "Updated create date of the artefact entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updatedAt",
            "in": "formData",
            "description": "Updated update date of artefact entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "thumbnail",
            "in": "formData",
            "description": "Updated image of the artefact in base64",
            "required": false,
            "type": "string",
            "format": "base64"
          },
          {
            "name": "coordX",
            "in": "formData",
            "description": "Updated x coordinate of the artefact in base64",
            "required": false,
            "type": "integer",
            "format": "base64"
          },
          {
            "name": "coordY",
            "in": "formData",
            "description": "Updated y coordinate of the artefact in base64",
            "required": false,
            "type": "integer",
            "format": "base64"
          },
          {
            "name": "zone",
            "in": "formData",
            "description": "Updated zone ID of artefact",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "artefacts"
        ],
        "summary": "Deletes a artefact",
        "description": "",
        "operationId": "deleteArtefact",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artefact id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Artefact not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/zones": {
      "get": {
        "tags": [
          "zones"
        ],
        "summary": "List all zones",
        "description": "",
        "operationId": "listZones",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "404": {
            "description": "Zones not found"
          }
        }
      },
      "post": {
        "tags": [
          "zones"
        ],
        "summary": "Add a new zone to the list",
        "description": "",
        "operationId": "addZones",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Zones that needs to be added to the list",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Zone"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/zones/{id}": {
      "get": {
        "tags": [
          "zones"
        ],
        "summary": "Find zone by ID",
        "description": "Returns a single zones",
        "operationId": "getZonesById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of zone to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Zone"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Zone not found"
          }
        }
      },
      "put": {
        "tags": [
          "zones"
        ],
        "summary": "Updates a zone with form data",
        "description": "",
        "operationId": "updateZonesWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "ID of the to be updated zone",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the zone",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "Updated description of the zone",
            "required": false,
            "type": "string"
          },
          {
            "name": "coordX",
            "in": "formData",
            "description": "Updated x coordinate of the artefact in base64",
            "required": false,
            "type": "integer",
            "format": "base64"
          },
          {
            "name": "coordY",
            "in": "formData",
            "description": "Updated y coordinate of the artefact in base64",
            "required": false,
            "type": "integer",
            "format": "base64"
          },
          {
            "name": "createdAt",
            "in": "formData",
            "description": "Updated create date of the zone entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updatedAt",
            "in": "formData",
            "description": "New update date of zone entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "zones"
        ],
        "summary": "Deletes a zone",
        "description": "",
        "operationId": "deleteZone",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Zone id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Zone not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/storeItems": {
      "get": {
        "tags": [
          "storeItems"
        ],
        "summary": "List all store items",
        "description": "",
        "operationId": "listStoreItems",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "404": {
            "description": "Store items not found"
          }
        }
      },
      "post": {
        "tags": [
          "storeItems"
        ],
        "summary": "Add a new store item to the list",
        "description": "",
        "operationId": "addStoreItem",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Store item that needs to be added to the list",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StoreItem"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/storeItems/{id}": {
      "get": {
        "tags": [
          "storeItems"
        ],
        "summary": "Find store item by ID",
        "description": "Returns a single store item",
        "operationId": "getStoreItemById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of store item to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/StoreItem"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Store item not found"
          }
        }
      },
      "put": {
        "tags": [
          "storeItems"
        ],
        "summary": "Updates a store item with form data",
        "description": "",
        "operationId": "updateStoreItemsWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "ID of the to be updated store item",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the store item",
            "required": true,
            "type": "string"
          },
          {
            "name": "cost",
            "in": "formData",
            "description": "Updated cost of the store item",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "inStock",
            "in": "formData",
            "description": "Updated stock status of the store item",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "createdAt",
            "in": "formData",
            "description": "Updated create date of the store item entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "updatedAt",
            "in": "formData",
            "description": "New update date of store item entry",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "thumbnail",
            "in": "formData",
            "description": "Updated thumbnail of the store item in base64",
            "required": false,
            "type": "string",
            "format": "base64"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "storeItems"
        ],
        "summary": "Deletes a store item",
        "description": "",
        "operationId": "deleteStoreItem",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store item id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Store item not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Artefact": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "acquisitionDate": {
          "type": "string",
          "format": "date-time"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "coordX": {
          "type": "number"
        },
        "coordY": {
          "type": "number"
        },
        "zoneId": {
          "type": "number"
        },
        "thumbnail": {
          "type": "string",
          "format": "base64"
        }
      },
      "xml": {
        "name": "Artefact"
      }
    },
    "Zone": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "Artefacts": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "acquisitionDate": {
                "type": "string",
                "format": "date-time"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "coordX": {
                "type": "number"
              },
              "coordY": {
                "type": "number"
              },
              "zoneId": {
                "type": "number"
              }
            }
          }
        },
        "Beacons": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "name": {
                "type": "string"
              },
              "visits": {
                "type": "integer",
                "format": "int64"
              },
              "macAddress": {
                "type": "string"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "coordX": {
                "type": "number"
              },
              "coordY": {
                "type": "number"
              },
              "zoneId": {
                "type": "number"
              }
            }
          }
        },
        "media": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "src": {
                "type": "string"
              },
              "type": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "video",
                    "image"
                  ]
                }
              },
              "artefactId": {
                "type": "number"
              }
            }
          }
        }
      },
      "xml": {
        "name": "Medias"
      }
    },
...
    "StoreItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "organiser": {
          "type": "string"
        },
        "cost": {
          "type": "integer",
          "format": "double"
        },
        "inStock": {
          "type": "boolean"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "thumbnail": {
          "type": "string",
          "format": "base64"
        }
      },
      "xml": {
        "name": "StoreItems"
      }
    },
...
  }
}